feat: refactoring file naming and imports
