+++ app/v2/core/infrastructure/file_system/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 
 
-from app.v2.core.infrastructure.file_system.main_file_writer_creator import FileWriterCreator
+from app.v2.core.infrastructure.file_system.file_writer_creator import FileWriterCreator
 
 
 __all__ = [
