--- app/v2/modules/file_manager/domain/services/create/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-
-from .file_writer_creator import CoreFileWriterCreator
-
-
-__all__ = [
-    "CoreFileWriterCreator",
-]
